Make room for a new trace function (#1351)
