+++ whitelist.txt
@@ -20,6 +20,7 @@ coincurve
 crypto
 E501
 EIP
+eip3155
 encodings
 endian
 eth
